requirements.txt
cheroot==6.0.0
CherryPy==13.1.0
cycler==0.10.0
DataSpyre==0.2.7
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.1
more-itertools==4.0.1
numpy==1.13.3
pandas==0.22.0
portend==2.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
tempora==1.10
